simplify thread function massively and remove heap usage
the thread callback function was allocating everything on the heap
without good reason, and constructing copies of strings manually
with combinations of malloc, strncpy, strncat, etc.
i simplified all such sites to use a single snprintf into a stack
buffer, which is then passed as argument to curl.

additionally i changed the while(queue->head) loop such that the
loop is always entered, but the queue head pointer is only fetched
when the lock is held, in order to prevent races with other threads.